+++ apps/owner/pages/index.tsx
@@ -3,7 +3,7 @@ import type { NextPage } from 'next'
 import Head from 'next/head'
 import Image from 'next/image'
 import { Button } from 'ui'
-import { Header } from '../Components/organism'
+import { Header } from '../components/organism'
 
 const Home: NextPage = () => (
   <div>
