+++ packages/ui/Themes/themeIndex.ts
@@ -1,0 +1,3 @@
+const Themes = require('./index')
+
+export { Themes }
